+++ src/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 const isFunction = fn => (typeof fn === 'function');
 const capitalize = str => (str.charAt(0).toUpperCase() + str.slice(1));
 
-export const composeWithState = initialState => BaseComponent => (
+export const composeWithState = (initialState, options = {}) => BaseComponent => (
   class extends React.PureComponent {
     constructor(props) {
       super(props);
@@ -12,10 +12,12 @@ export const composeWithState = initialState => BaseComponent => (
       } else {
         this.state = initialState;
       }
+      const { setters = {} } = options;
       this.state = this.state || {};
       this.stateSetters = {};
       Object.keys(this.state).forEach((key) => {
-        this.stateSetters[`set${capitalize(key)}`] = (val) => {
+        const setter = setters[key] || `set${capitalize(key)}`;
+        this.stateSetters[setter] = (val) => {
           this.setState({ [key]: val });
         };
       });